requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
dnspython>=2.6.0
python-whois>=0.9.4
tldextract>=5.1.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
